Split public API contract types by resource
